+++ lib/scheduler.ml
@@ -51,13 +51,9 @@ module type StreamDescriptor = sig
 
   val flush_write_body : t -> max_bytes:int -> int
 
-  val on_more_output_available : t -> (unit -> unit) -> unit
-
   val finish_stream : t -> Stream.closed_reason -> unit
 
   val is_idle : t -> bool
-
-  val closed : t -> Stream.closed option
 end
 
 module Make (Streamd : StreamDescriptor) = struct
@@ -95,6 +91,8 @@ module Make (Streamd : StreamDescriptor) = struct
           ; mutable flow : Settings.WindowSize.t
                 (* inbound flow control, what the client is allowed to send. *)
           ; mutable inflow : Settings.WindowSize.t
+          ; mutable marked_for_removal :
+              (Stream_identifier.t * Stream.closed) list
           }
           -> root node
       | Stream :
@@ -142,6 +140,7 @@ module Make (Streamd : StreamDescriptor) = struct
       ; all_streams = StreamsTbl.create ~random:true capacity
       ; flow = Settings.WindowSize.default_initial_window_size
       ; inflow = Settings.WindowSize.default_initial_window_size
+      ; marked_for_removal = []
       }
 
   let create ~parent ~initial_window_size descriptor =
@@ -320,10 +319,6 @@ module Make (Streamd : StreamDescriptor) = struct
   let allowed_to_receive (Connection root) (Stream stream) size =
     size < root.inflow && size < stream.inflow
 
-  let requires_output t stream =
-    let (Stream { descriptor; _ }) = stream in
-    allowed_to_transmit t stream && Streamd.requires_output descriptor
-
   let write (Connection root as t) stream_node =
     let (Stream ({ descriptor; _ } as stream)) = stream_node in
     (* From RFC7540§6.9.1:
@@ -357,6 +352,26 @@ module Make (Streamd : StreamDescriptor) = struct
     in
     stream.t <- tlast_p + (n * 256 / stream.priority.weight)
 
+  let mark_for_removal (Connection root) id closed =
+    root.marked_for_removal <- (id, closed) :: root.marked_for_removal
+
+  let implicitly_close_idle_stream descriptor max_seen_ids =
+    let implicitly_close_stream descriptor =
+      if Streamd.is_idle descriptor then
+        (* From RFC7540§5.1.1:
+         *   The first use of a new stream identifier implicitly closes all
+         *   streams in the "idle" state that might have been initiated by
+         *   that peer with a lower-valued stream identifier. *)
+        Streamd.finish_stream descriptor Finished
+    in
+    let max_client_stream_id, max_pushed_stream_id = max_seen_ids in
+    let stream_id = Streamd.id descriptor in
+    if Stream_identifier.is_request stream_id then (
+      if stream_id < max_client_stream_id then
+        implicitly_close_stream descriptor)
+    else if stream_id < max_pushed_stream_id then
+      implicitly_close_stream descriptor
+
   (* Scheduling algorithm from https://goo.gl/3sSHXJ (based on nghttp2):
    *
    * 1  def schedule(p):
@@ -373,7 +388,7 @@ module Make (Streamd : StreamDescriptor) = struct
    * 12
    * 13 schedule(0)
    *)
-  let flush t =
+  let flush t max_seen_ids =
     let rec schedule : type a. a node -> int * bool = function
       | Connection p ->
         (* The root can never send data. *)
@@ -384,10 +399,11 @@ module Make (Streamd : StreamDescriptor) = struct
           if subtree_is_active then (
             update_t i_node written;
             p.children <- PriorityQueue.add id i_node children')
-          else
+          else (
+            implicitly_close_idle_stream i.descriptor max_seen_ids;
             (* XXX(anmonteiro): we may not want to remove from the tree right
              * away. *)
-            p.children <- children';
+            p.children <- children');
           written, subtree_is_active
         | None ->
           (* Queue is empty, see line 6 above. *)
@@ -410,14 +426,34 @@ module Make (Streamd : StreamDescriptor) = struct
             if subtree_is_active then (
               update_t i_node written;
               p.children <- PriorityQueue.add id i_node children')
-            else
-              p.children <- children';
+            else (
+              implicitly_close_idle_stream i.descriptor max_seen_ids;
+              p.children <- children');
             written, subtree_is_active
           | None ->
             (* Queue is empty, see line 6 above. *)
             0, false)
     in
-    ignore (schedule t)
+    let (Connection root) = t in
+    ignore (schedule t);
+    root.marked_for_removal
+    <- List.fold_left
+         (fun acc (id, closed) ->
+           (* When a stream completes, i.e. doesn't require more output and
+            * enters the `Closed` state, we set a TTL value which represents
+            * the * number of writer yields that the stream has before it is
+            * removed * from the connection Hash Table. By doing this we avoid
+            * losing some * potentially useful information regarding the
+            * stream's state at the * cost of keeping it around for a little
+            * while longer. *)
+           if closed.Stream.ttl == 0 then (
+             StreamsTbl.remove root.all_streams id;
+             acc)
+           else (
+             closed.ttl <- closed.ttl - 1;
+             (id, closed) :: acc))
+         []
+         root.marked_for_removal
 
   let check_flow flow growth flow' =
     (* Check for overflow on 32-bit systems. *)
@@ -461,44 +497,6 @@ module Make (Streamd : StreamDescriptor) = struct
     | Stream ({ inflow; _ } as stream) ->
       stream.inflow <- inflow - size
 
-  let on_more_output_available t (max_client_stream_id, max_pushed_stream_id) k
-    =
-    let (Connection root) = t in
-    let implicitly_close_stream descriptor =
-      if Streamd.is_idle descriptor then
-        (* From RFC7540§5.1.1:
-         *   The first use of a new stream identifier implicitly closes all
-         *   streams in the "idle" state that might have been initiated by
-         *   that peer with a lower-valued stream identifier. *)
-        Streamd.finish_stream descriptor Finished
-      else
-        match Streamd.closed descriptor with
-        | Some c ->
-          (* When a stream completes, i.e. doesn't require more output and
-           * enters the `Closed` state, we set a TTL value which represents the
-           * number of writer yields that the stream has before it is removed
-           * from the connection Hash Table. By doing this we avoid losing some
-           * potentially useful information regarding the stream's state at the
-           * cost of keeping it around for a little while longer. *)
-          if c.ttl = 0 then
-            StreamsTbl.remove root.all_streams (Streamd.id descriptor)
-          else
-            c.ttl <- c.ttl - 1
-        | None ->
-          ()
-    in
-    StreamsTbl.iter
-      (fun id stream_node ->
-        let (Stream { descriptor; _ }) = stream_node in
-        if Stream_identifier.is_request id then (
-          if id < max_client_stream_id then
-            implicitly_close_stream descriptor)
-        else if id < max_pushed_stream_id then
-          implicitly_close_stream descriptor;
-        if requires_output t stream_node then
-          Streamd.on_more_output_available descriptor k)
-      root.all_streams
-
   let pp_hum fmt t =
     let rec pp_hum_inner level fmt t =
       let pp_binding fmt (i, Stream { children; t; _ }) =
@@ -515,27 +513,3 @@ module Make (Streamd : StreamDescriptor) = struct
     in
     pp_hum_inner 0 fmt t
 end
-
-module Client_scheduler = Make (struct
-  include Stream
-
-  type t = Respd.t
-
-  let flush_write_body = Respd.flush_request_body
-
-  let requires_output = Respd.requires_output
-
-  let on_more_output_available = Respd.on_more_output_available
-end)
-
-module Server_scheduler = Make (struct
-  include Stream
-
-  type t = Reqd.t
-
-  let flush_write_body = Reqd.flush_response_body
-
-  let requires_output = Reqd.requires_output
-
-  let on_more_output_available = Reqd.on_more_output_available
-end)