+++ lib/server_connection.ml
@@ -35,7 +35,16 @@
 module AB = Angstrom.Buffered
 module Reader = Parse.Reader
 module Writer = Serialize.Writer
-module Scheduler = Scheduler.Server_scheduler
+
+module Scheduler = Scheduler.Make (struct
+  include Stream
+
+  type t = Reqd.t
+
+  let flush_write_body = Reqd.flush_response_body
+
+  let requires_output = Reqd.requires_output
+end)
 
 type request_handler = Reqd.t -> unit
 
@@ -109,13 +118,19 @@ let shutdown_writer t =
   if not (is_active t) then
     wakeup_writer t
 
-(* let error_code t = if is_active t then Reqd.error_code (current_reqd_exn t)
-   else None *)
-
 let shutdown t =
   shutdown_reader t;
   shutdown_writer t;
   wakeup_writer t
+
+(* Handling frames against closed streams is hard. See:
+ * https://docs.google.com/presentation/d/1iG_U2bKTc9CnKr0jPTrNfmxyLufx_cK2nNh9VjrKH6s
+ *)
+let was_closed_or_implicitly_closed t stream_id =
+  if Stream_identifier.is_request stream_id then
+    Stream_identifier.(stream_id <= t.max_client_stream_id)
+  else
+    Stream_identifier.(stream_id <= t.max_pushed_stream_id)
 
 let report_error t = function
   | Error.ConnectionError (error, data) ->
@@ -152,11 +167,12 @@ let report_error t = function
     | Some reqd ->
       Stream.reset_stream reqd error
     | None ->
-      (* Possible if the stream was going to enter the Idle state (first time
-       * we saw e.g. a PRIORITY frame for it) but had e.g. a
-       * FRAME_SIZE_ERROR. *)
-      let frame_info = Writer.make_frame_info stream_id in
-      Writer.write_rst_stream t.writer frame_info error);
+      if not (was_closed_or_implicitly_closed t stream_id) then
+        (* Possible if the stream was going to enter the Idle state (first time
+         * we saw e.g. a PRIORITY frame for it) but had e.g. a
+         * FRAME_SIZE_ERROR. *)
+        let frame_info = Writer.make_frame_info stream_id in
+        Writer.write_rst_stream t.writer frame_info error);
     wakeup_writer t
 
 let report_connection_error t ?(additional_debug_data = "") error =
@@ -176,8 +192,14 @@ let report_exn t exn =
     let additional_debug_data = Printexc.to_string exn in
     report_connection_error t ~additional_debug_data Error.InternalError
 
-let on_stream_closed t () =
-  t.current_client_streams <- t.current_client_streams - 1
+let on_close_stream t id ~active closed =
+  if active then
+    (* From RFC7540§5.1.2:
+     *   Streams that are in the "open" state or in either of the "half-closed"
+     *   states count toward the maximum number of streams that an endpoint is
+     *   permitted to open. *)
+    t.current_client_streams <- t.current_client_streams - 1;
+  Scheduler.mark_for_removal t.streams id closed
 
 let create_push_stream ({ max_pushed_stream_id; _ } as t) () =
   if not t.settings.enable_push then
@@ -205,14 +227,14 @@ let create_push_stream ({ max_pushed_stream_id; _ } as t) () =
         ~max_frame_size:t.settings.max_frame_size
         t.writer
         t.error_handler
-        (on_stream_closed t)
+        (on_close_stream t pushed_stream_id)
     in
     Scheduler.add
       t.streams (* TODO: *)
                 (* ?priority *)
       ~initial_window_size:t.settings.initial_window_size
       reqd;
-    Ok (reqd, fun () -> wakeup_writer t)
+    Ok reqd
 
 let handle_headers t ~end_stream reqd active_stream headers =
   (* From RFC7540§5.1.2:
@@ -273,7 +295,9 @@ let handle_headers t ~end_stream reqd active_stream headers =
                  * value for [request_body_buffer_size]. *)
                 t.config.request_body_buffer_size
             in
-            Body.create (Bigstringaf.create buffer_size)
+            Body.create
+              (Bigstringaf.create buffer_size)
+              active_stream.wakeup_writer
         in
         let request_info = Reqd.create_active_request request request_body in
         if end_stream then (
@@ -385,7 +409,7 @@ let open_stream t frame_header ?priority headers_block =
             ~max_frame_size:t.settings.max_frame_size
             t.writer
             t.error_handler
-            (on_stream_closed t)
+            (on_close_stream t stream_id)
         in
         Scheduler.add
           t.streams
@@ -427,6 +451,7 @@ let open_stream t frame_header ?priority headers_block =
       Reqd.create_active_stream
         t.hpack_encoder
         t.config.response_body_buffer_size
+        (fun () -> wakeup_writer t)
         (create_push_stream t)
     in
     reqd.state <- Active (Open (PartialHeaders partial_headers), active_stream);
@@ -676,11 +701,12 @@ let process_data_frame t { Frame.frame_header; _ } bstr =
          *   stream error (Section 5.4.2) of type STREAM_CLOSED. *)
         report_stream_error t stream_id Error.StreamClosed)
     | None ->
-      (* From RFC7540§5.1:
-       *   idle: [...] Receiving any frame other than HEADERS or PRIORITY on
-       *   a stream in this state MUST be treated as a connection error
-       *   (Section 5.4.1) of type PROTOCOL_ERROR. *)
-      report_connection_error t Error.ProtocolError)
+      if not (was_closed_or_implicitly_closed t stream_id) then
+        (* From RFC7540§5.1:
+         *   idle: [...] Receiving any frame other than HEADERS or PRIORITY on
+         *   a stream in this state MUST be treated as a connection error
+         *   (Section 5.4.1) of type PROTOCOL_ERROR. *)
+        report_connection_error t Error.ProtocolError)
 
 let process_priority_frame t { Frame.frame_header; _ } priority =
   let { Frame.stream_id; _ } = frame_header in
@@ -710,16 +736,16 @@ let process_priority_frame t { Frame.frame_header; _ } priority =
        * Note:
        *   if we're receiving a PRIORITY frame for a stream that we already
        *   removed from the tree (i.e. can't be found in the hash table, and
-       *   for which the stream ID is strictly smaller than the maximum stream
+       *   for which the stream ID is smaller then or equal to the max stream
        *   id that the client has opened), don't bother processing it. *)
-      if Stream_identifier.(stream_id >= t.max_client_stream_id) then
+      if not (was_closed_or_implicitly_closed t stream_id) then
         let reqd =
           Stream.create
             stream_id
             ~max_frame_size:t.settings.max_frame_size
             t.writer
             t.error_handler
-            (on_stream_closed t)
+            (on_close_stream t stream_id)
         in
         Scheduler.add
           t.streams
@@ -766,9 +792,9 @@ let process_rst_stream_frame t { Frame.frame_header; _ } error_code =
         Stream.finish_stream reqd (ResetByThem error_code))
     | None ->
       (* We might have removed the stream from the hash table. If its stream
-       * id is strictly smaller than the max client stream id we've seen, then
-       * it must have been closed. *)
-      if Stream_identifier.(stream_id >= t.max_client_stream_id) then
+       * id is smaller than or equal to the max client stream id we've seen,
+       * then it must have been closed. *)
+      if not (was_closed_or_implicitly_closed t stream_id) then
         (* From RFC7540§6.4:
          *   RST_STREAM frames MUST NOT be sent for a stream in the "idle"
          *   state. If a RST_STREAM frame identifying an idle stream is
@@ -987,11 +1013,12 @@ let process_window_update_frame t { Frame.frame_header; _ } window_increment =
          *   error (Section 5.4.1) of type PROTOCOL_ERROR. *)
         ())
     | None ->
-      (* From RFC7540§5.1:
-       *   idle: [...] Receiving any frame other than HEADERS or PRIORITY on
-       *   a stream in this state MUST be treated as a connection error
-       *   (Section 5.4.1) of type PROTOCOL_ERROR. *)
-      report_connection_error t Error.ProtocolError
+      if not (was_closed_or_implicitly_closed t stream_id) then
+        (* From RFC7540§5.1:
+         *   idle: [...] Receiving any frame other than HEADERS or PRIORITY on
+         *   a stream in this state MUST be treated as a connection error
+         *   (Section 5.4.1) of type PROTOCOL_ERROR. *)
+        report_connection_error t Error.ProtocolError
 
 let process_continuation_frame t { Frame.frame_header; _ } headers_block =
   let { Frame.stream_id; flags; _ } = frame_header in
@@ -1216,27 +1243,16 @@ let read t bs ~off ~len =
 let read_eof t bs ~off ~len =
   Reader.read_with_more (reader t) bs ~off ~len Complete
 
-let flush_response_body t =
-  if is_active t then
-    Scheduler.flush t.streams
-
 let next_write_operation t =
-  flush_response_body t;
+  if is_active t then
+    Scheduler.flush t.streams (t.max_client_stream_id, t.max_pushed_stream_id);
   Writer.next t.writer
 
 let report_write_result t result = Writer.report_result t.writer result
 
 let yield_writer t k =
-  if is_active t then (
-    let { max_client_stream_id; max_pushed_stream_id; _ } = t in
-    Scheduler.on_more_output_available
-      t.streams
-      (max_client_stream_id, max_pushed_stream_id)
-      k;
-    (* We need to let both the connection and each individual streams wake up
-     * the connection, given that some control frames can be scheduled to be
-     * written when all the streams are yielding. *)
-    on_wakeup_writer t k)
+  if is_active t then
+    on_wakeup_writer t k
   else if Writer.is_closed t.writer then
     k ()
   else